+++ Caravan.DataAccess/Sql/PostgreSql/PostgreSqlDataSourceManager.cs
@@ -1,6 +1,8 @@
-﻿using System.Data.Common;
+﻿using Dapper;
 using Finsa.Caravan.DataAccess.Core;
 using Npgsql;
+using System;
+using System.Data.Common;
 
 namespace Finsa.Caravan.DataAccess.Sql.PostgreSql
 {
@@ -9,6 +11,18 @@ namespace Finsa.Caravan.DataAccess.Sql.PostgreSql
         public override CaravanDataSourceKind DataSourceKind
         {
             get { return CaravanDataSourceKind.PostgreSql; }
+        }
+
+        public override DateTime DataSourceDateTime
+        {
+            get
+            {
+                using (var connection = OpenConnection())
+                {
+                    const string query = "TODO";
+                    return connection.ExecuteScalar<DateTime>(query);
+                }
+            }
         }
 
         public override string ElaborateConnectionString(string connectionString)