+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.49.0 (10/03/2016) ###
+* [UPD] HttpProxyMiddleware ora ignora i cookie e l'header "Host".
+* [UPD] Il servizio "help/serviceInfo" ora restituisce data e ora del server e della sorgente dati.
+
 ### v1.48.0 (08/03/2016) ###
 
 * [REM] Il pacchetto EFInteractiveViews non è più necessario.